Merge. En archivo.sql se crea la tabla solo si no existe. Endpoint POST predicciones se termino y ahora los errores se atajan desde app.py. Agregado de mysql-connector-python en requirements.

+++ db.py
@@ -46,6 +46,10 @@ def delete_partido_por_id(id):
     cursor = coneccion.cursor(dictionary=True)
     
     try:
+        cursor.execute('SELECT id FROM partidos WHERE id = %s', (id_ingresado,))
+        partido = cursor.fetchone()
+        if not partido:
+            return False
         cursor.execute('DELETE FROM partidos WHERE id = %s', (id_ingresado,))
         coneccion.commit()
         return True
@@ -76,7 +80,7 @@ def agregar_resultado_por_id(id, goles_local, goles_visitante):
         if not partido:
             return False
         else:
-            cursor.execute('INSERT INTO resultados (partido_id, goles_visitante, goles_local) VALUES(%s, %s, %s) ON DUPLICATE KEY UPDATE goles_visitante = VALUES(goles_visitante), goles_local = VALUES(goles_local)', (id_ingresado, goles_visitante, goles_local,))
+            cursor.execute('UPDATE partidos SET goles_local = %s, goles_visitante = %s WHERE id = %s', (goles_local, goles_visitante, id_ingresado,))
             coneccion.commit()
             return True
     finally:
@@ -129,25 +133,10 @@ def hacer_prediccion(usuario_id, partido_id, goles_local, goles_visitante):
     cursor = coneccion.cursor(dictionary=True)
 
     try:
-        # Buscar partido
-        cursor.execute("SELECT id FROM partidos WHERE id = %s", (partido_id,))
-        partido = cursor.fetchone()
-        if not partido:
-            return False
-        
-        # Buscar usuario
-        cursor.execute("SELECT id FROM usuarios WHERE id = %s", (usuario_id,))
-        usuario = cursor.fetchone()
-        if not usuario:
-            return False
-        
-        # cursor.execute("SELECT  goles_visitante goles_local FROM resultados WHERE goles_visitante IS NULL or goles_local IS NULL")
-        # se refiere a resultados, pero primero hay que hacer el cambio de los goles, agregarlos a la tabla partidos.
-
-        cursor.execute("SELECT usuario_id FROM predicciones WHERE usuario_id = %s AND partido_id = %s", (usuario_id, partido_id))
+        cursor.execute("SELECT id FROM predicciones WHERE usuario_id = %s AND partido_id = %s", (usuario_id, partido_id,))
         prediccion_hecha = cursor.fetchone()
         if prediccion_hecha:
-            return False, 'No se puede hacer mas de una prediccion por partido.', 400   #esto esta para probar, no sabemos si se puede devolver todo esto.
+            return False
 
         cursor.execute('INSERT INTO predicciones (usuario_id, partido_id, goles_local, goles_visitante) VALUES(%s, %s, %s, %s)', (usuario_id, partido_id, goles_local, goles_visitante))
         coneccion.commit()
